prefer 'files' over .npmignore
Explicitly include *only* the 'lib' directory in the published package. Compared
to the .npmignore file this now excludes the .npmignore and .babelrc files, as
well as the tests. The tests require the src modules and dev dependencies so
they wouldn't have worked anyway. The other files are unnecessary when
publishing.